Re-use code instead of duplicating it.
